twitter_collect_mentioned/requirements.txt
schedule
psycopg2-binary
requests-oauthlib